Move them (again) to local components folder
