+++ src/routes/(app)/components/side-nav/GoalTree.svelte
