+++ src/routes/(app)/components/side-nav/NavItem.svelte
